chore(lint): clippy lint policy + fix real lints (workspace clean)
- [workspace.lints]: deny clippy::all, allow a curated set of pre-existing
  style/complexity/perf lints (manual_div_ceil — not even clippy will auto-fix it
  due to overflow semantics — needless_range_loop, too_many_arguments, etc.).
  Hoist unexpected_cfgs(stacks_dump) to [workspace.lints.rust]. Each crate opts in
  via [lints] workspace = true.
- Fixed the REAL ones the policy surfaced (not allowed): self_assignment +
  identity_op (Metal prefill), redundant_guards (cli), unnecessary_lazy_evaluations
  + let_and_return (server). clippy --fix applied the machine-applicable remainder.
- cargo clippy --workspace --release is now 0 errors; cargo fmt --all clean.

+++ crates/lumen-convert/src/arch/qwen35_moe.rs
@@ -552,7 +552,7 @@ fn compute_layer_shape_qwen35moe(
                 n_elements % 32 == 0,
                 "Q4_0 requires elements divisible by 32, got {n_elements} for {name}"
             );
-            let q4_size = (n_elements as u64 / 32) * 18;
+            let q4_size = (n_elements / 32) * 18;
             let slice = TensorSlice {
                 offset: *blob_offset,
                 length: q4_size,
